requirements.txt
Django==1.11.1
Pillow==4.1.1
boto==2.47.0
celery==4.0.2
dj-database-url==0.4.2
django-filter==1.0.4
django-storages==1.5.2
django-mptt==0.8.6
django-versatileimagefield==1.7.0
djangorestframework-jwt==1.10.0
djangorestframework==3.6.3
drf-nested-routers==0.90.0
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
raven==6.1.0
redis==2.10.5
requests==2.17.3
whitenoise==3.3.0
git+git://github.com/Brogency/djangorestframework-camel-case@a6a448360941130678b62aebed5d5d09ba1214bb#egg=djangorestframework-camel-case

ipdb
ipdbplugin
django-extensions==1.7.9
ipython==6.1.0

django-nose==1.4.4
rednose==1.2.2
Faker==0.7.12
factory_boy==2.8.1
coverage==4.4.1

